Removed some parts
