--- src/com/wire/collab/FEdit.java
@@ -1,27 +1,0 @@
-/**
- * @name Simple Java NotePad
- * @package ph.notepad
- * @file UI.java
- * @author SORIA Pierre-Henry
- * @email pierrehs@hotmail.com
- * @link http://github.com/pH-7
- * @copyright Copyright Pierre-Henry SORIA, All Rights Reserved.
- * @license Apache (http://www.apache.org/licenses/LICENSE-2.0)
- * @create 2012-05-04
- * @update 2015-09-4
- * @modifiedby Achintha Gunasekara
- * @modweb http://www.achinthagunasekara.com
- * @modemail contact@achinthagunasekara.com
- */
-
-package com.wire.collab;
-
-import javax.swing.*;
-
-public class FEdit {
-
-    public static void clear(JTextArea textArea) {
-        textArea.setText("");
-    }
-
-}
