--- src/com/wire/collab/HighlightText.java
@@ -1,43 +1,0 @@
-package com.wire.collab;
-
-import javax.swing.text.*;
-import java.awt.*;
-
-public class HighlightText extends DefaultHighlighter.DefaultHighlightPainter {
-
-    public HighlightText(Color color) {
-        super(color);
-    }
-
-    public void highLight(JTextComponent textComp, String[] pattern) {
-        removeHighlights(textComp);
-
-        try {
-            Highlighter hilite = textComp.getHighlighter();
-            Document doc = textComp.getDocument();
-            String text = doc.getText(0, doc.getLength());
-            for (int i = 0; i < pattern.length; i++) {
-                int pos = 0;
-
-                while ((pos = text.indexOf(pattern[i], pos)) >= 0) {
-                    hilite.addHighlight(pos, pos + pattern[i].length(), this);
-                    pos += pattern[i].length();
-                }
-            }
-        } catch (BadLocationException e) {
-        }
-
-    }
-
-    public void removeHighlights(JTextComponent textComp) {
-
-        Highlighter hilite = textComp.getHighlighter();
-        Highlighter.Highlight[] hilites = hilite.getHighlights();
-
-        for (int i = 0; i < hilites.length; i++) {
-            if (hilites[i].getPainter() instanceof HighlightText) {
-                hilite.removeHighlight(hilites[i]);
-            }
-        }
-    }
-}
